Move classes out of `describer-manager` into their own files.
